reworked file manger to use relative paths. added clip managing UI

+++ DiscordBot/Models/Internal/Configs/Blazor.cs
@@ -5,5 +5,5 @@ public sealed class Blazor
     /// <summary>
     /// Defaults to 100MB, can be overwritten via configuration.
     /// </summary>
-    public int FileSizeLimitMb { get; set; } = 100;
+    public long FileSizeLimitMb { get; set; } = 100;
 }
